Add LLasso score CLI
