+++ constants.py
@@ -1,0 +1,15 @@
+"""
+Stores all used API keys and emails
+"""
+
+OPENAI_API = ""
+LANGCHAIN_API = ""
+EMAIL = ""
+TOGETHER_API = ""
+OPEN_ROUTER = ""
+OMIM_KEYS = [
+    ""
+]
+
+OMIM_PERSIST_DIRECTORY = "DATABASE DIRECTORY"  # Vector store path
+OMIM_DATA_FILE = "DATABASE DIRECTORY"  # Path to JSON data file
