+++ pyproject.toml
@@ -33,6 +33,8 @@ dependencies = [
     "openai==1.60.0",
     "requests==2.32.3",
     "scikit-learn==1.6.1",
+    "transformers==4.49.0",
+    "xmltodict==0.14.2",
 ]
 
 [tool.setuptools.packages.find]
