--- src/constants.py
@@ -1,12 +1,0 @@
-"""
-Stores all used API keys and emails
-"""
-
-APIKEY = ""
-LANGCHAIN_API = ""
-EMAIL = ""
-TOGETHER_API = ""
-OPEN_ROUTER = ""
-OMIM_KEYS = [
-    ""
-]
